feat(server files): reorganize server-side files - move index.js into server folder - remove unused environment variables in .env.default - move config folder into server - add all required environment variables - break out test/index.js file into several files based on routes - update connection string in server/db/index.js [Delivers #158686613]
