+++ .env.default
@@ -1,10 +1,6 @@
 NODE_ENV=
 
-DB_HOST=
-DB_USER=
-DB_PORT=
-DB_PASSWORD=
-DB_NAME=
-DATABASE_URL_DEV=
-DATABASE_URL_TEST=
-DATABASE_URL_PRO=
+CONN_STRING_DEV=
+CONN_STRING_PRO=
+CONN_STRING_TEST=
+SECRET_KEY=
