--- config/config.js
@@ -1,25 +1,0 @@
-require('dotenv').config();
-
-module.exports = {
-  development: {
-    username: process.env.DB_USER,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-    host: process.env.DB_HOST,
-    port: process.env.DB_PORT,
-    dialect: 'postgres',
-    url: process.env.DATABASE_URL_DEV
-  },
-  test: {
-    username: process.env.DB_USERNAME,
-    password: process.env.DB_PASSWORD,
-    database: process.env.DBNAME,
-    host: process.env.DB_HOST,
-    port: process.env.DB_PORT,
-    dialect: 'postgres',
-    url: process.env.DATABASE_URL_TEST
-  },
-  production: {
-    url: process.env.DATABASE_URL_PRO
-  }
-};
